Update foreign keys to the unsigned integer data type
MySQL was colliding on foreign key creation due to unmatched data types.

+++ database/migrations/2018_08_04_185018_create_member_loans_table.php
@@ -16,13 +16,13 @@
       Schema::create('member_loans', function (Blueprint $table) {
         $table->increments('member_loan_id');
         
-        $table->integer('member_id');
-        $table->smallInteger('loan_type_id');
+        $table->unsignedInteger('member_id');
+        $table->unsignedInteger('loan_type_id');
         $table->string('loan_purpose');
         $table->decimal('loan_amount');
         $table->integer('repayment_period');
-        $table->smallInteger('loan_repayment_status_id');
-        $table->smallInteger('loan_issuing_status_id');
+        $table->unsignedInteger('loan_repayment_status_id');
+        $table->unsignedInteger('loan_issuing_status_id');
         
         $table->timestamps();
